# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DND AI Roleplay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/characters": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "List Characters",
        "operationId": "list_characters_api_v1_characters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterSummary"
                  },
                  "type": "array",
                  "title": "Response List Characters Api V1 Characters Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Create Character",
        "operationId": "create_character_api_v1_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Get Character",
        "operationId": "get_character_api_v1_characters__character_id__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "characters"
        ],
        "summary": "Update Character",
        "operationId": "update_character_api_v1_characters__character_id__patch",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/avatar": {
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Upload Avatar",
        "operationId": "upload_avatar_api_v1_characters__character_id__avatar_post",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_characters__character_id__avatar_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/skills": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Get Skills",
        "operationId": "get_skills_api_v1_characters__character_id__skills_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSetView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "characters"
        ],
        "summary": "Update Skills",
        "operationId": "update_skills_api_v1_characters__character_id__skills_put",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillSetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSetView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/memories": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "List Memories",
        "operationId": "list_memories_api_v1_characters__character_id__memories_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoryView"
                  },
                  "title": "Response List Memories Api V1 Characters  Character Id  Memories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Create Memory",
        "operationId": "create_memory_api_v1_characters__character_id__memories_post",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/memories/{memory_id}": {
      "patch": {
        "tags": [
          "characters"
        ],
        "summary": "Update Memory",
        "operationId": "update_memory_api_v1_characters__character_id__memories__memory_id__patch",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "characters"
        ],
        "summary": "Delete Memory",
        "operationId": "delete_memory_api_v1_characters__character_id__memories__memory_id__delete",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}:export": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Export Character",
        "operationId": "export_character_api_v1_characters__character_id__export_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/sheet-versions:preview": {
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Preview Character Sheet",
        "operationId": "preview_character_sheet_api_v1_characters__character_id__sheet_versions_preview_post",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_character_sheet_api_v1_characters__character_id__sheet_versions_preview_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/sheet-versions/{version_id}:activate": {
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Activate Character Sheet",
        "operationId": "activate_character_sheet_api_v1_characters__character_id__sheet_versions__version_id__activate_post",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetActivation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_preview_character_sheet_api_v1_characters__character_id__sheet_versions_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_character_sheet_api_v1_characters__character_id__sheet_versions_preview_post"
      },
      "Body_upload_avatar_api_v1_characters__character_id__avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_v1_characters__character_id__avatar_post"
      },
...
      "CharacterCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "roleplayPrompt": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Roleplayprompt"
          },
          "voiceSamples": {
            "type": "string",
            "maxLength": 20000,
            "title": "Voicesamples",
            "default": ""
          },
          "narrationNotes": {
            "type": "string",
            "maxLength": 4000,
            "title": "Narrationnotes",
            "default": ""
          },
          "maxHp": {
            "type": "integer",
            "maximum": 100000.0,
            "exclusiveMinimum": 0.0,
            "title": "Maxhp"
          }
        },
        "type": "object",
        "required": [
          "name",
          "roleplayPrompt",
          "maxHp"
        ],
        "title": "CharacterCreate"
      },
      "CharacterDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "maxHp": {
            "type": "integer",
            "title": "Maxhp"
          },
          "avatarPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarpath"
          },
          "hasRoleplayPrompt": {
            "type": "boolean",
            "title": "Hasroleplayprompt"
          },
          "hasActiveSheet": {
            "type": "boolean",
            "title": "Hasactivesheet"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "roleplayPrompt": {
            "type": "string",
            "title": "Roleplayprompt"
          },
          "voiceSamples": {
            "type": "string",
            "title": "Voicesamples"
          },
          "narrationNotes": {
            "type": "string",
            "title": "Narrationnotes"
          },
          "profileContent": {
            "type": "string",
            "title": "Profilecontent"
          },
          "profileStatus": {
            "$ref": "#/components/schemas/ProfileStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "maxHp",
          "avatarPath",
          "hasRoleplayPrompt",
          "hasActiveSheet",
          "revision",
          "createdAt",
          "updatedAt",
          "roleplayPrompt",
          "voiceSamples",
          "profileContent",
          "profileStatus",
          "narrationNotes"
        ],
        "title": "CharacterDetail"
      },
...
      "CharacterSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "maxHp": {
            "type": "integer",
            "title": "Maxhp"
          },
          "avatarPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarpath"
          },
          "hasRoleplayPrompt": {
            "type": "boolean",
            "title": "Hasroleplayprompt"
          },
          "hasActiveSheet": {
            "type": "boolean",
            "title": "Hasactivesheet"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "maxHp",
          "avatarPath",
          "hasRoleplayPrompt",
          "hasActiveSheet",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "CharacterSummary"
      },
      "CharacterUpdate": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "roleplayPrompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Roleplayprompt"
          },
          "voiceSamples": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Voicesamples"
          },
          "narrationNotes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrationnotes"
          },
          "maxHp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxhp"
          },
          "profileContent": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000
              },
              {
                "type": "null"
              }
            ],
            "title": "Profilecontent"
          }
        },
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "CharacterUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MemoryCreate"
      },
      "MemoryOrigin": {
        "type": "string",
        "enum": [
          "DM",
          "AUTO"
        ],
        "title": "MemoryOrigin"
      },
      "MemoryUpdate": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          }
        },
        "type": "object",
        "title": "MemoryUpdate"
      },
      "MemoryView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "characterId": {
            "type": "string",
            "title": "Characterid"
          },
          "sourceCampaignId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcecampaignid"
          },
          "sourceSessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcesessionid"
          },
          "origin": {
            "$ref": "#/components/schemas/MemoryOrigin"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "characterId",
          "sourceCampaignId",
          "sourceSessionId",
          "origin",
          "content",
          "pinned",
          "createdAt",
          "updatedAt"
        ],
        "title": "MemoryView"
      },
...
      "ProfileStatus": {
        "type": "string",
        "enum": [
          "READY",
          "NEEDS_REBUILD",
          "GENERATING",
          "FAILED"
        ],
        "title": "ProfileStatus"
      },
...
      "SkillSetUpdate": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "modifiers": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/SkillName"
            },
            "type": "object",
            "title": "Modifiers"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "modifiers"
        ],
        "title": "SkillSetUpdate"
      },
      "SkillSetView": {
        "properties": {
          "characterId": {
            "type": "string",
            "title": "Characterid"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "modifiers": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/SkillName"
            },
            "type": "object",
            "title": "Modifiers"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "characterId",
          "revision",
          "modifiers",
          "updatedAt"
        ],
        "title": "SkillSetView"
      },
...
    }
  }
}